# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cloud Edge Backend API",
    "version": "2.4.0",
    "description": "S1-S6 workflow backend and UI-10 management: execution snapshots, guarded logical deletion, ADMIN/USER accounts, image upload/build/distribution, deployments and managed resources. One execution chain; no throughput SDK, long-horizon DQN, physical purge or full Kestra feature parity."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18085"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
  "paths": {
...
    "/api/namespaces/{namespace}/me": {
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "currentUser",
        "description": "Enabled human account in this namespace. CONNECT machines denied.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or namespace/action denied"
          },
          "404": {
            "description": "Active record not found"
          },
          "409": {
            "description": "Referenced/active record, revision or identity conflict"
          },
          "422": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/api/namespaces/{namespace}/me/password": {
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "put": {
        "operationId": "changeOwnPassword",
        "description": "Verify current password; BCrypt new password. Old credentials stop authenticating; frontend logs out. Bootstrap config never overwrites changes.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or namespace/action denied"
          },
          "404": {
            "description": "Active record not found"
          },
          "409": {
            "description": "Referenced/active record, revision or identity conflict"
          },
          "422": {
            "description": "Invalid request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              }
            }
          }
        }
      }
    },
    "/api/namespaces/{namespace}/users": {
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "listUsers",
        "description": "ADMIN only. Full target namespace scope must be within administrator scope; no password hashes.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfile"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or namespace/action denied"
          },
          "404": {
            "description": "Active record not found"
          },
          "409": {
            "description": "Referenced/active record, revision or identity conflict"
          },
          "422": {
            "description": "Invalid request"
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ]
      },
      "post": {
        "operationId": "createUser",
        "description": "ADMIN only within own namespaces. ADMIN and USER both have READ/WRITE/EXECUTE. No CONNECT creation/custom roles.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or namespace/action denied"
          },
          "404": {
            "description": "Active record not found"
          },
          "409": {
            "description": "Referenced/active record, revision or identity conflict"
          },
          "422": {
            "description": "Invalid request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        }
      }
    },
    "/api/namespaces/{namespace}/users/{name}": {
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "name",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "put": {
        "operationId": "updateUser",
        "description": "ADMIN only within full target scope. Update enabled/role/namespaces; cannot disable/demote yourself. No account deletion.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or namespace/action denied"
          },
          "404": {
            "description": "Active record not found"
          },
          "409": {
            "description": "Referenced/active record, revision or identity conflict"
          },
          "422": {
            "description": "Invalid request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        }
      }
    },
    "/api/namespaces/{namespace}/users/{name}/password": {
      "parameters": [
        {
          "name": "namespace",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "name",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "put": {
        "operationId": "resetUserPassword",
        "description": "ADMIN only, other human accounts within full namespace scope. Own password uses personal center.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Role or namespace/action denied"
          },
          "404": {
            "description": "Active record not found"
          },
          "409": {
            "description": "Referenced/active record, revision or identity conflict"
          },
          "422": {
            "description": "Invalid request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordReset"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
      "UserProfile": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "namespaces": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "type": "string"
            }
          },
          "actions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "READ",
                "WRITE",
                "EXECUTE"
              ]
            }
          }
        },
        "required": [
          "name",
          "role",
          "enabled",
          "namespaces",
          "actions"
        ],
        "additionalProperties": false
      },
      "UserCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[A-Za-z][A-Za-z0-9_.-]{0,99}$"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 10,
            "description": "Maximum 72 UTF-8 bytes; never returned"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER"
            ]
          },
          "namespaces": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "password",
          "role",
          "namespaces"
        ],
        "additionalProperties": false
      },
      "UserUpdate": {"type":"object","properties":{"role":{"type":"string","enum":["ADMIN","USER"]},"enabled":{"type":"boolean"},"namespaces":{"type":"array","minItems":1,"uniqueItems":true,"items":{"type":"string"}}},"required":["role","enabled","namespaces"],"additionalProperties":false},
      "PasswordChange": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "format": "password"
          },
          "newPassword": {
            "type": "string",
            "format": "password",
            "minLength": 10,
            "description": "Maximum 72 UTF-8 bytes; never returned"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "additionalProperties": false
      },
      "PasswordReset": {"type":"object","properties":{"password":{"type":"string","format":"password","minLength":10,"description":"Maximum 72 UTF-8 bytes; never returned"}},"required":["password"],"additionalProperties":false},
...
    }
  }
}